requirements.txt
azure-mgmt-resource
azure-identity
msrestazure